Added pseudocode for the challenge algorithm
Most of the hard work should already be done, we just need to pass a few new variables and spin the sonar. I'll be around tonight and  most of tomorrow (but not in College) if this raises any questions. Also, Mike and Malhar, please join the group on CATE so that I can submit the dummy file.